[GDI32] - Some fixes for CreateDIBitmap (Patch by Victor Martinez, modified by me) CORE-6420 #resolve - Fix warning on MSVC
svn path=/trunk/; revision=57440

+++ reactos/win32ss/gdi/gdi32/objects/efloat.c
@@ -27,7 +27,7 @@ EFtoF(EFLOAT_S * efp)
     Exp = efp->lExp;
     Sign = SIGN(Mant);
 
-    if (Sign) Mant = -Mant;
+    if (Sign) Mant = -(LONG)Mant;
     Mant >>= 7;
     Exp += (EXCESS-1);
